resources, resource_types, resource_groups
Added all the files for resources, resource_types, and resource_groups.
Note that some editing is still required, because I did not know what to
rename the word "calendar" to in many files (after making a copy of a
resource_calendar file).  Example:  I don't know what to change
"$edit_calendar_rules" to.